+++ SWLOR.Game.Server/Feature/RecipeDefinition/EngineeringRecipeDefinition/StarshipRecipes.cs
@@ -47,26 +47,6 @@ namespace SWLOR.Game.Server.Feature.RecipeDefinition.EngineeringRecipeDefinition
                 .Component("aluminum", 3)
                 .Component("fiberp_ruined", 3)
                 .Component("elec_ruined", 3);
-
-            // Ship Missiles x5
-            _builder.Create(RecipeType.Missile5, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_missile")
-                .Level(5)
-                .Quantity(5)
-                .RequirementPerk(PerkType.StarshipBlueprints, 1)
-                .Component("ref_tilarium", 1)
-                .Component("elec_ruined", 1);
-
-            // Hypermatter Capsule x1
-            _builder.Create(RecipeType.FuelCapsule1, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_fuelcapsule")
-                .Level(8)
-                .Quantity(1)
-                .RequirementPerk(PerkType.StarshipBlueprints, 1)
-                .Component("ref_tilarium", 3)
-                .Component("elec_ruined", 1);
         }
 
         private void Tier2()
@@ -96,16 +76,6 @@ namespace SWLOR.Game.Server.Feature.RecipeDefinition.EngineeringRecipeDefinition
                 .Component("steel", 3)
                 .Component("fiberp_flawed", 3)
                 .Component("elec_flawed", 3);
-
-            // Ship Missiles x15
-            _builder.Create(RecipeType.Missile15, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_missile")
-                .Level(15)
-                .Quantity(15)
-                .RequirementPerk(PerkType.StarshipBlueprints, 2)
-                .Component("ref_currian", 5)
-                .Component("elec_flawed", 1);
         }
 
         private void Tier3()
@@ -135,26 +105,6 @@ namespace SWLOR.Game.Server.Feature.RecipeDefinition.EngineeringRecipeDefinition
                 .Component("obsidian", 3)
                 .Component("fiberp_good", 3)
                 .Component("elec_good", 3);
-
-            // Ship Missiles x25
-            _builder.Create(RecipeType.Missile25, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_missile")
-                .Level(25)
-                .Quantity(25)
-                .RequirementPerk(PerkType.StarshipBlueprints, 3)
-                .Component("ref_idailia", 7)
-                .Component("elec_good", 2);
-
-            // Hypermatter Capsule x3
-            _builder.Create(RecipeType.FuelCapsule3, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_fuelcapsule")
-                .Level(28)
-                .Quantity(3)
-                .RequirementPerk(PerkType.StarshipBlueprints, 1)
-                .Component("ref_idailia", 3)
-                .Component("elec_good", 1);
         }
 
         private void Tier4()
@@ -184,16 +134,6 @@ namespace SWLOR.Game.Server.Feature.RecipeDefinition.EngineeringRecipeDefinition
                 .Component("crystal", 3)
                 .Component("fiberp_imperfect", 3)
                 .Component("elec_imperfect", 3);
-
-            // Ship Missiles x35
-            _builder.Create(RecipeType.Missile35, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_missile")
-                .Level(25)
-                .Quantity(35)
-                .RequirementPerk(PerkType.StarshipBlueprints, 4)
-                .Component("ref_barinium", 9)
-                .Component("elec_imperfect", 2);
         }
 
         private void Tier5()
@@ -278,37 +218,6 @@ namespace SWLOR.Game.Server.Feature.RecipeDefinition.EngineeringRecipeDefinition
                 .Component("diamond", 3)
                 .Component("fiberp_high", 3)
                 .Component("elec_high", 3);
-
-            // Ship Missiles x45
-            _builder.Create(RecipeType.Missile45, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_missile")
-                .Level(45)
-                .Quantity(45)
-                .RequirementPerk(PerkType.StarshipBlueprints, 5)
-                .Component("ref_gostian", 11)
-                .Component("elec_high", 3);
-
-            // Proton Bomb x5
-            _builder.Create(RecipeType.ProtonBomb, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("proton_bomb")
-                .Level(45)
-                .Quantity(5)
-                .RequirementPerk(PerkType.StarshipBlueprints, 5)
-                .Component("ref_gostian", 3)
-                .Component("elec_high", 2)
-                .Component("ref_arda", 1);
-
-            // Hypermatter Capsule x5
-            _builder.Create(RecipeType.FuelCapsule5, SkillType.Engineering)
-                .Category(RecipeCategoryType.Starship)
-                .Resref("ship_fuelcapsule")
-                .Level(48)
-                .Quantity(5)
-                .RequirementPerk(PerkType.StarshipBlueprints, 5)
-                .Component("ref_gostian", 3)
-                .Component("elec_high", 1);
 
             // Republic Thranta
             _builder.Create(RecipeType.CorvetteRepThranta, SkillType.Engineering)
